refactor: Server Side, Client Side 분리
